requirements.txt
amqp==5.3.1
anyio==4.12.0
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==6.0
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-graphql-jwt==0.4.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.7
graphql-relay==3.2.0
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
kombu==5.6.1
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
promise==2.3
prompt_toolkit==3.0.52
psycopg2==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
